+++ lib/ContextParser.ts
@@ -4,7 +4,7 @@ import {ERROR_CODES, ErrorCoded} from "./ErrorCoded";
 import {FetchDocumentLoader} from "./FetchDocumentLoader";
 import {IDocumentLoader} from "./IDocumentLoader";
 import {IJsonLdContext, IJsonLdContextNormalizedRaw, IPrefixValue, JsonLdContext} from "./JsonLdContext";
-import {JsonLdContextNormalized} from "./JsonLdContextNormalized";
+import {JsonLdContextNormalized, defaultExpandOptions, IExpandOptions} from "./JsonLdContextNormalized";
 import {Util} from "./Util";
 
 // tslint:disable-next-line:no-var-requires
@@ -981,24 +981,3 @@ export interface IParseOptions {
   ignoreScopedContexts?: boolean;
 }
 
-export interface IExpandOptions {
-  /**
-   * If compact IRI prefixes can end with any kind of character in simple term definitions,
-   * instead of only the default gen-delim characters (:,/,?,#,[,],@).
-   */
-  allowPrefixNonGenDelims: boolean;
-  /**
-   * If compact IRI prefixes ending with a non-gen-delim character
-   * can be forced as a prefix using @prefix: true.
-   */
-  allowPrefixForcing: boolean;
-  /**
-   * If @vocab values are allowed contain IRIs relative to @base.
-   */
-  allowVocabRelativeToBase: boolean;
-}
-export const defaultExpandOptions: IExpandOptions = {
-  allowPrefixForcing: true,
-  allowPrefixNonGenDelims: false,
-  allowVocabRelativeToBase: true,
-};
